+++ src/main/java/com/ltonetwork/client/core/transacton/CancelLease.java
@@ -3,15 +3,14 @@ package com.ltonetwork.client.core.transacton;
 import com.google.common.primitives.Bytes;
 import com.google.common.primitives.Longs;
 import com.ltonetwork.client.exceptions.BadMethodCallException;
-import com.ltonetwork.client.utils.Encoder;
 import com.ltonetwork.client.utils.JsonObject;
 
 import java.nio.charset.StandardCharsets;
 
 public class CancelLease extends Transaction {
     private final static long MINIMUM_FEE = 100_000_000;
-    private final static int TYPE = 9;
-    private final static int VERSION = 2;
+    private final static byte TYPE = 9;
+    private final static byte VERSION = 2;
     private final long leaseId;
 //    private final Lease lease;
