+++ src/main/java/com/ltonetwork/client/core/transacton/Lease.java
@@ -12,8 +12,8 @@ import java.nio.charset.StandardCharsets;
 
 public class Lease extends Transaction {
     private final static long MINIMUM_FEE = 100_000_000;
-    private final static int TYPE = 8;
-    private final static int VERSION = 2;
+    private final static byte TYPE = 8;
+    private final static byte VERSION = 2;
     private final long amount;
     private final Address recipient;
